MVP - Moves components to their own folders
