+++ packages/orchestrator-ui-components/src/components/WFOWorkflowSteps/index.ts
@@ -1,2 +1,5 @@
 export * from './WFOStep';
+export * from './WFOStepList';
+export * from './WFOStepsRawJson';
+export * from './WFOStepStatusIcon';
 export * from './WFOWorkflowStepList';
